requirements.txt
altair==5.3.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
comtypes==1.4.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.51.0
fqdn==1.5.1
gensim==4.3.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==2.10
imbalanced-learn==0.13.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.26.4
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pickle-mixin==1.0.2
pillow==10.3.0
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.22
pydeck==0.9.0
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.2
pypiwin32==223
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyttsx3==2.90
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
setuptools==76.0.0
six==1.16.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soccerdata==1.8.7
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.34.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
undetected-chromedriver==3.5.5
unicode==2.9
Unidecode==1.3.8
uri-template==1.3.0
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.10
wordcloud==1.9.3
wrapt==1.17.2
wsproto==1.2.0
